+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to **bUnit** will be documented in this file. The project ad
 ## [Unreleased]
 
 ### Added
-- Added support for `RendererInfo` (`.net9.0`).
+- Added support for `RendererInfo` and `AssignedRenderMode` (`.net9.0`).
 
 ## [1.36.0] - 2024-11-12